# docs/openapi/spec.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "All the information needed to talk to the API.\n\nTo raise see anything wrong? Raise an [issue](https://github.com/mailchain/mailchain/issues)",
    "title": "Mailchain API",
    "version": "0.0.1"
  },
  "basePath": "/api",
  "paths": {
...
    "/messages": {
      "get": {
        "description": "Check the protocol, network, address combination for Mailchain messages.",
        "tags": [
          "Messages"
        ],
        "summary": "Get Mailchain messages.",
        "operationId": "GetMessages",
        "parameters": [
          {
            "pattern": "0x[a-fA-F0-9]{40}",
            "type": "string",
            "example": "0xde0b295669a9fd93d5f28d9ec85e40f4cb697bae",
            "x-go-name": "Address",
            "description": "Address to use when looking for messages.",
            "name": "address",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "mainnet",
              "goerli",
              "ropsten",
              "rinkeby",
              "local"
            ],
            "type": "string",
            "example": "goerli",
            "x-go-name": "Network",
            "description": "Network to use when looking for messages.",
            "name": "network",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "ethereum"
            ],
            "type": "string",
            "example": "ethereum",
            "x-go-name": "Protocol",
            "description": "Protocol to use when looking for messages.",
            "name": "protocol",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetMessagesResponse"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      },
      "post": {
        "description": "Securely send message on the protocol and network specified in the query string to the address.\nOnly the private key holder for the recipient address can decrypted any encrypted contents.\n\nCreate mailchain message\nEncrypt content with public key\nStore message\nEncrypt location\nStore encrypted location on the blockchain.",
        "tags": [
          "Send"
        ],
        "summary": "Send message.",
        "operationId": "SendMessage",
        "parameters": [
          {
            "enum": [
              "mainnet",
              "goerli",
              "ropsten",
              "rinkeby",
              "local"
            ],
            "type": "string",
            "example": "goerli",
            "x-go-name": "Network",
            "description": "Network to use when sending a message.",
            "name": "network",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "ethereum"
            ],
            "type": "string",
            "example": "ethereum",
            "x-go-name": "Protocol",
            "description": "Protocol to use when sending a message.",
            "name": "protocol",
            "in": "query",
            "required": true
          },
          {
            "description": "Message to send",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SendMessageRequestBody"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/StatusOK"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "GetMessagesResponseHeaders": {
      "type": "object",
      "properties": {
        "date": {
          "description": "When the message was created, this can be different to the transaction data of the message.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Date",
          "readOnly": true,
          "example": "12 Mar 19 20:23 UTC"
        },
        "from": {
          "description": "The sender of the message",
          "type": "string",
          "x-go-name": "From",
          "readOnly": true,
          "example": "Charlotte \u003c5602ea95540bee46d03ba335eed6f49d117eab95c8ab8b71bae2cdd1e564a761@ropsten.ethereum\u003e"
        },
        "message-id": {
          "description": "Unique identifier of the message",
          "type": "string",
          "x-go-name": "MessageID",
          "readOnly": true,
          "example": "47eca011e32b52c71005ad8a8f75e1b44c92c99fd12e43bccfe571e3c2d13d2e9a826a550f5ff63b247af471@mailchain"
        },
        "reply-to": {
          "description": "Reply to if the reply address is different to the from address.",
          "type": "string",
          "x-go-name": "ReplyTo",
          "readOnly": true
        },
        "to": {
          "description": "The recipient of the message",
          "type": "string",
          "x-go-name": "To",
          "readOnly": true
        }
      },
      "x-go-name": "getHeaders",
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
    "GetMessagesResponseMessage": {
      "type": "object",
      "properties": {
        "body": {
          "description": "Body of the mail message",
          "type": "string",
          "x-go-name": "Body",
          "readOnly": true,
          "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur maximus metus ante, sit amet ullamcorper dui hendrerit ac."
        },
        "headers": {
          "$ref": "#/definitions/GetMessagesResponseHeaders"
        },
        "read": {
          "description": "Read status of the message",
          "type": "boolean",
          "x-go-name": "Read",
          "readOnly": true,
          "example": true
        },
        "status": {
          "type": "string",
          "x-go-name": "Status",
          "readOnly": true
        },
        "status-code": {
          "type": "string",
          "x-go-name": "StatusCode",
          "readOnly": true
        },
        "subject": {
          "description": "Subject of the mail message",
          "type": "string",
          "x-go-name": "Subject",
          "readOnly": true,
          "example": "Hello world"
        }
      },
      "x-go-name": "getMessage",
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
...
    "PostMessagesResponseHeaders": {
      "type": "object",
      "required": [
        "from",
        "to"
      ],
      "properties": {
        "from": {
          "description": "The sender of the message",
          "type": "string",
          "x-go-name": "From",
          "example": "Charlotte \u003c5602ea95540bee46d03ba335eed6f49d117eab95c8ab8b71bae2cdd1e564a761@ropsten.ethereum\u003e"
        },
        "reply-to": {
          "description": "Reply to if the reply address is different to the from address.",
          "type": "string",
          "x-go-name": "ReplyTo"
        },
        "to": {
          "description": "The recipient of the message",
          "type": "string",
          "x-go-name": "To"
        }
      },
      "x-go-name": "PostHeaders",
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
    "PostMessagesResponseMessage": {
      "type": "object",
      "required": [
        "headers",
        "body",
        "subject",
        "public-key"
      ],
      "properties": {
        "body": {
          "description": "Body of the mail message",
          "type": "string",
          "x-go-name": "Body",
          "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur maximus metus ante,"
        },
        "headers": {
          "$ref": "#/definitions/PostMessagesResponseHeaders"
        },
        "public-key": {
          "description": "Public key of the recipient to encrypt with",
          "type": "string",
          "x-go-name": "PublicKey"
        },
        "subject": {
          "description": "Subject of the mail message",
          "type": "string",
          "x-go-name": "Subject",
          "example": "Hello world"
        }
      },
      "x-go-name": "PostMessage",
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
    "SendMessageRequestBody": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "$ref": "#/definitions/PostMessagesResponseMessage"
        }
      },
      "x-go-name": "PostRequestBody",
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    }
  },
  "responses": {
...
    "GetMessagesResponse": {
      "description": "GetResponse Holds the response messages",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/GetMessagesResponseMessage"
        }
      }
    },
...
    "NotFoundError": {
      "description": "Description of an error.",
      "headers": {
        "code": {
          "type": "string",
          "example": "404",
          "description": "Code describing the error"
        },
        "message": {
          "type": "string",
          "example": "Not found.",
          "description": "Description of the error"
        }
      }
    },
    "StatusOK": {
      "description": "StatusOK Description of an StatusOK."
    },
    "ValidationError": {
      "description": "Description of an error.",
      "headers": {
        "code": {
          "type": "string",
          "example": "422",
          "description": "Code describing the error"
        },
        "message": {
          "type": "string",
          "example": "Response to invalid input",
          "description": "Description of the error"
        }
      }
    }
  }
}